remove modal and vue.js, create guests only by email
