+++ appinfo/routes.php
@@ -55,4 +55,11 @@ return [
 			'verb' => 'PUT'
 		]
 	],
+	'ocs' => [
+		[
+			'name' => 'sharees#search',
+			'url' => '/api/v1/sharees',
+			'verb' => 'GET',
+		]
+	]
 ];
