--- package.json
@@ -1,40 +1,0 @@
-{
-  "name": "guests",
-  "version": "0.4.1",
-  "description": "Allow classifying users or contacts as guests.",
-  "main": "src/default.js",
-  "directories": {
-    "test": "tests"
-  },
-  "scripts": {
-    "dev": "webpack  --progress --colors --watch",
-    "build": "webpack -p"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/owncloud/guests.git"
-  },
-  "keywords": [
-    "share",
-    "guest",
-    "user"
-  ],
-  "author": "Felix Heidecke, Ilja Neumann",
-  "license": "AGPL-3.0",
-  "bugs": {
-    "url": "https://github.com/owncloud/guests/issues"
-  },
-  "homepage": "https://github.com/owncloud/guests#readme",
-  "dependencies": {
-    "babel-core": "^6.24.1",
-    "babel-loader": "^7.0.0",
-    "babel-preset-es2015": "^6.24.1",
-    "css-loader": "^0.28.0",
-    "less": "^2.7.2",
-    "less-loader": "^4.0.3",
-    "style-loader": "^0.16.1",
-    "uglify-js": "git+https://github.com/mishoo/UglifyJS2.git#harmony",
-    "vue": "^2.2.6",
-    "webpack": "^2.4.1"
-  }
-}
