--- src/styles/default.less
@@ -1,70 +1,0 @@
-// Use #app-guest to establish a namespace
-
-#app-guests {
-
-  @modal-width: 500px;
-  @modal-height: 360px;
-  @modal-gutter: 20px;
-
-  @moda-z-index: 500;
-
-  .modal {
-    box-sizing: border-box;
-    position: absolute;
-    z-index: @moda-z-index + 1;
-    left: ~"calc(50% - "(@modal-width / 2)~")";
-    top: 20%;
-
-    width: @modal-width;
-    background: white;
-    padding: @modal-gutter;
-
-    &-title,
-    &-body {
-      margin-bottom: @modal-gutter * 2;
-    }
-
-    .form {
-      &-group {
-        margin: @modal-gutter / 2 0;
-      }
-
-      &-label,
-      &-input {
-        box-sizing: border-box;
-        display: block;
-        width: 100%;
-        transition:
-          background .333s ease-out,
-          border .333s ease-out;
-
-        &._error {
-          background: fade(red, 10%);
-          border: 1px solid fade(red, 20%);
-          color: red;
-        }
-      }
-
-    }
-
-    &-backdrop {
-      position: absolute;
-      top: 0;
-      left: 0;
-      right: 0;
-      bottom: 0;
-      background-color: rgba(0, 0, 0, .5);
-      z-index: @moda-z-index;
-      content: '';
-      width: 100vw;
-      height: 100vh;
-      display: block;
-    }
-  }
-
-  // Mods
-
-  &.-animate-to {
-    opacity: 1;
-  }
-}
